removal of neo wrappers as not needed
