+++ spynnaker8/utilities/__init__.py
@@ -1,6 +1,4 @@
 from .data_holder import DataHolder
 from .id import ID
-from .spynnaker8_neo_block import SpynnakerNeoBlock
-from .spynnaker8_neo_segment import SpynnakerNeoSegment
 
-__all__ = ["DataHolder", "ID", "SpynnakerNeoBlock", "SpynnakerNeoSegment"]
+__all__ = ["DataHolder", "ID"]
